converter: new 'select' function
The purpose of this function is to tag all RLum.Analysis and RLum.Data.Curve as selected, or to de-select them. The general idea is to keep the complete input data set and all further functions work on the object returned by get_selected.
This function is still in alpha stage and not yet implemented in the application itself.