abanico: warn user on logged statistics, see R-Lum/Luminescence#50
By default, the abanico plot operates on the logged DE values and also calculates the summary statistics on the logged DE values. While not wrong per sé (there are statistical reasons), it is usually not what the users expects and which leaves them wondering if the calculations are wrong. This help text is to avoid confusion why e.g. the mean differs from the "normal" mean of the DE values they might have calculated beforehand.